Group incentive gets external system boolean

+++ db/migrations/202101111106_add_from_ext_sys_to_group_incentive.rb
@@ -1,13 +1,21 @@
 Sequel.migration do
   up do
-    alter_table(:group_incentives) do
-      add_column :from_external_system, :boolean, default: false
+    if ENV['CLIENT_CODE'] != 'kr'
+      puts 'Migration not run (only applicable to Kromco)'
+    else
+      alter_table(:group_incentives) do
+        add_column :from_external_system, :boolean, default: false
+      end
     end
   end
 
   down do
-    alter_table(:group_incentives) do
-      drop_column :from_external_system
+    if ENV['CLIENT_CODE'] != 'kr'
+      puts 'Migration not run (only applicable to Kromco)'
+    else
+      alter_table(:group_incentives) do
+        drop_column :from_external_system
+      end
     end
   end
 end
